finding ordinal value of elements
The Python ord() function returns the Unicode value from a given character, this Unicode value is represented as an integer. This particular Python function only accepts a string with the length of one, it then returns the Unicode integer (decimal) value that represents the given string.  
The Unicode Standard* is the universal character encoding standard used for representation of text for computer processing.